refactor - Extract the stacktrace parsing to util
